@@ -29,16 +29,16 @@
   },
   "homepage": "https://github.com/insin/react-octicon",
   "dependencies": {
-    "octicons": "^3.1.0"
+    "octicons": "4.x"
   },
   "peerDependencies": {
     "react": ">=0.13.0"
   },
   "devDependencies": {
-    "eslint-config-jonnybuchanan": "^2.0.1",
-    "nwb": "~0.2.0",
-    "react": "^0.14.3",
-    "react-addons-test-utils": "^0.14.3",
-    "react-dom": "^0.14.3"
+    "copy-to-clipboard": "3.x",
+    "eslint-config-jonnybuchanan": "5.x",
+    "nwb": "0.20.x",
+    "react": "16.x",
+    "react-dom": "16.x"
   }
 }
